requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.1
cron-descriptor==1.4.5
Django==4.2.23
django-celery-beat==2.8.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.3
django-ratelimit==4.1.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==35.2.2
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==5.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
django-redis
